+++ pkgs/intl/lib/find_locale.dart
@@ -1,7 +1,3 @@
-// Copyright (c) 2023, the Dart project authors. Please see the AUTHORS file
-// for details. All rights reserved. Use of this source code is governed by a
-// BSD-style license that can be found in the LICENSE file.
-
 export 'src/intl_default.dart' // Stub implementation
     // Browser implementation
     if (dart.library.html) 'intl_browser.dart'