+++ pkgs/intl/CHANGELOG.md
@@ -6,6 +6,7 @@
  * Update to CLDR v43.
  * Switch QAR currency name to Riyal.
  * Add license headers to some files.
+ * Update CVE currency symbol.
  * Fix issue [#483](https://github.com/dart-lang/i18n/issues/483) about date parsing with a `yy` skeleton.
 
 ## 0.18.1